Add display.py, and move all display functionality from task_stats.py to display.py.
